Tracking down errors in bolide simulations. Simulations crash after some time, uncertain cause.

+++ src/fvm/fvm_bolide_restart_driver.cpp
@@ -479,7 +479,7 @@ void FVMBolideRestartDriver::restart(Job* job) {
                                 break;
                             case 2:
                                 //porosity
-                                fluid_body->n_e(i) = std::stoi(line);
+                                fluid_body->n_e(i) = std::stod(line);
                                 break;
                             default:
                                 //do nothing
